Move OGGBundle specific ConstructorSection to opengever.bundle
