+++ opengever/bundle/cfgs/oggbundle.cfg
@@ -36,7 +36,7 @@ blueprint = opengever.bundle.bundlesource
 blueprint = opengever.setup.resolveguid
 
 [constructor]
-blueprint = opengever.setup.constructor
+blueprint = opengever.bundle.constructor
 
 [fileloader]
 blueprint = opengever.bundle.fileloader
