+++ opengever/bundle/tests/test_section_constructor.py
@@ -1,8 +1,8 @@
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
+from opengever.bundle.sections.constructor import ConstructorSection
+from opengever.bundle.sections.constructor import InvalidType
 from opengever.bundle.tests import MockTransmogrifier
-from opengever.setup.sections.constructor import ConstructorSection
-from opengever.setup.sections.constructor import InvalidType
 from opengever.testing import FunctionalTestCase
 from plone import api
 from zope.interface.verify import verifyClass
