+++ opengever/bundle/transmogrifier.zcml
@@ -19,6 +19,11 @@
       />
 
   <utility
+      component=".sections.constructor.ConstructorSection"
+      name="opengever.bundle.constructor"
+    />
+
+  <utility
       component=".sections.fileloader.FileLoaderSection"
       name="opengever.bundle.fileloader"
     />
