+++ opengever/setup/transmogrifier.zcml
@@ -80,9 +80,4 @@
     name="opengever.setup.resolveguid"
     />
 
-  <utility
-    component=".sections.constructor.ConstructorSection"
-    name="opengever.setup.constructor"
-    />
-
 </configure>
